--- src/main/java/wooteco/subway/domain/PathGenerator.java
@@ -1,8 +1,0 @@
-package wooteco.subway.domain;
-
-import java.util.List;
-
-public interface PathGenerator {
-
-    Path findPath(List<Section> sections, Station from, Station to);
-}
