+++ src/main/java/wooteco/subway/service/dto/PathResponse.java
@@ -2,8 +2,8 @@ package wooteco.subway.service.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import wooteco.subway.domain.Fare;
-import wooteco.subway.domain.Path;
+import wooteco.subway.domain.fare.Fare;
+import wooteco.subway.domain.path.Path;
 
 public class PathResponse {
     private final List<StationResponse> stations;
