+++ src/test/java/wooteco/subway/domain/farestrategy/AgeStrategyTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import wooteco.subway.domain.fare.farestrategy.AgeStrategy;
 
 class AgeStrategyTest {
